==> 2_desc_PCA.Rmd <==

```{r setup}
source("dependencies.R")
```

```{r}
approach <- 'replic_830'

if (grepl("830", approach)) {
  data2004 <- read_parquet("data/data2004_830.parquet")
} else if (grepl("1077", approach)) {
  data2004 <- read_parquet("data/data2004_1077.parquet")
} else if (grepl("2015", approach)) {
  data2004 <- read_parquet("data/data2004_1077.parquet")}
```

```{r}
q2_vars <- c('clseusa')
q3_vars <- c('ambornin', 'amcit', 'amlived', 'amenglsh', 'amchrstn', 'amgovt', 'amfeel')
q4_vars <- c('amcitizn', 'amshamed', 'belikeus', 'ambetter', 'ifwrong')
q5_vars = c('proudsss', 'proudgrp', 'proudpol', 'prouddem', 'proudeco', 'proudspt', 'proudart', 'proudhis', 'proudmil', 'proudsci')

if (approach == 'own_1215') {
  q2_vars <- c('clsetown', 'clsestat', 'clseusa', 'clsenoam')
  q3_vars <- c(q3_vars, 'amancstr')
  q4_vars <- c(q4_vars, 'amsports', 'lessprd')
}

q2_vars_f <- paste0(q2_vars, "_f")
q3_vars_f <- paste0(q3_vars, "_f")
q4_vars_f <- paste0(q4_vars, "_f")
q5_vars_f <- paste0(q5_vars, "_f")

q2_vars_n <- paste0(q2_vars, "_n")
q3_vars_n <- paste0(q3_vars, "_n")
q4_vars_n <- paste0(q4_vars, "_n")
q5_vars_n <- paste0(q5_vars, "_n")

f_outcomes <- c(q2_vars_f, q3_vars_f, q4_vars_f, q5_vars_f)
n_outcomes <- c(q2_vars_n, q3_vars_n, q4_vars_n, q5_vars_n)
```

# Outcomes

## Likert plots

```{r}
likert_data <- data2004 %>%
  set_variable_labels(
    clseusa_f = 'America') %>%
  set_variable_labels(
    ambornin_f = 'To have been born in America',
    amcit_f = 'To have American citizenship',
    amlived_f = 'To have lived in America for most of one\'s life',
    amenglsh_f = 'To be able to speak English',
    amchrstn_f = 'To be a Christian',
    amgovt_f = 'To respect America\'s political institutions and laws',
    amfeel_f = 'To feel American') %>%
  set_variable_labels(
    amcitizn_f = 'I would rather be a citizen of America than of any other country in the world',
    belikeus_f = 'The world would be a better place if people from other countries were more like the Americans',
    ambetter_f = 'Generally speaking, America is a better country than most other countries',
    ifwrong_f = 'People should support their country even if their country is in the wrong',
    amshamed_f = 'There are some things about America today that make me feel ashamed of America') %>%
  set_variable_labels(
    prouddem_f = 'The way democracy works',
    proudpol_f = 'Its political influence in the world',
    proudeco_f = 'America\'s economic achievements',
    proudsss_f = 'Its social security system',
    proudsci_f = 'Its scientific and technological achievements',
    proudspt_f = 'Its achievements in sports',
    proudart_f = 'Its achievements in the arts & literature',
    proudmil_f = 'America\'s armed forces',
    proudhis_f = 'Its history',
    proudgrp_f = 'Its fair and equal treatment of all groups in society')

if (approach == 'own_12015') {
  likert_data <- likert_data %>%
    set_variable_labels(
      clsetown_f = 'Your town or city',
      clsestat_f = 'Your state',
      clsenoam_f = 'North America') %>%
    set_variable_labels(
      amancstr_f = 'To have American ancestry') %>%
    set_variable_labels(
      amsports_f = 'When my country does well in international sports, it makes me proud to be American',
      lessprd_f = 'I am often less proud of America than I would like to be')
}

full_levels <- c("Not at all", "Not very much", "Neutral", "Somewhat / fairly", "Very much")
likert_data[f_outcomes] <- lapply(f_outcomes, function(var) {
  x <- likert_data[[var]]
  lbl <- var_label(x)
  x <- factor(x, levels = full_levels)
  var_label(x) <- lbl
  x
})
```

```{r}
plot <- gglikert(likert_data,
                 include = all_of(f_outcomes),
                 sort = "ascending") +
  scale_fill_brewer(palette = "RdYlBu")

ggsave("output/desc/likert_plot_2004.png", plot = plot, width = 10, height = 8, dpi = 300)
```

```{r}
p1 <- gglikert(likert_data, include = all_of(q2_vars_f), sort = "ascending", add_totals = FALSE) +
  scale_x_continuous(labels = function(x) abs(x)*100, limits = c(-0.51, 1)) + 
  scale_fill_brewer(palette = "RdYlBu") + 
  ggtitle("How close do you feel to?") +
  theme(legend.position = "none",
        plot.title = element_text(hjust = 0.5, face = "bold"))

p2 <- gglikert(likert_data, include = all_of(q3_vars_f), sort = "ascending", add_totals = FALSE) +
  scale_x_continuous(labels = function(x) abs(x)*100, limits = c(-0.51, 1)) + 
  scale_fill_brewer(palette = "RdYlBu") +
  ggtitle("How important do you think each of the following is?") +
  theme(legend.position = "none",
        plot.title = element_text(hjust = 0.5, face = "bold"))

p3 <- gglikert(likert_data, include = all_of(q4_vars_f), sort = "ascending", add_totals = FALSE) + 
  scale_x_continuous(labels = function(x) abs(x)*100, limits = c(-0.51, 1)) +
  scale_fill_brewer(palette = "RdYlBu") +
  ggtitle("How much do you agree or disagree with the following statements?") +
  theme(plot.title = element_text(hjust = 0.5, face = "bold"))

p4 <- gglikert(likert_data, include = all_of(q5_vars_f), sort = "ascending", add_totals = FALSE) + 
  scale_x_continuous(labels = function(x) abs(x)*100, limits = c(-0.51, 1)) +
  scale_fill_brewer(palette = "RdYlBu") +
  ggtitle("How proud are you of America in each of the following?") +
  theme(legend.position = "none",
        plot.title = element_text(hjust = 0.5, face = "bold"))

plot <- ggpubr::ggarrange(p1, p2, p3, p4, 
                          nrow=4, 
                          ncol=1,
                          common.legend = TRUE,
                          legend = "bottom",
                          align = "hv",
                          heights = c(
                            max(0.22, length(q2_vars_f)*0.11),
                            max(0.22, length(q3_vars_f)*0.11),
                            max(0.22, length(q4_vars_f)*0.11),
                            max(0.22, length(q5_vars_f)*0.11)))

ggsave("output/desc/likert_plots_2004.png", plot = plot, width = 10, height = 15, dpi = 300)
```

## PCA

Il faut bien décider / justifier de scaler ou pas (scale.unit = TRUE / FALSE). 

- Sans scaling : 2 axes, portant 26 et 13 % de l'inertie totale, dont l'opposition est essentiellement tirée par 3 variables (amchrstn, ambornin, amlived) 

- Avec scaling : 2 axes, potant 26 et 11% de l'inertie totale, dont l'opposition est tirée par une quinzaine de variables (dont les précédentes, dont l'effet est en quelque sorte dilué)

```{r}
f_controls <- c('sex_f', 'race_f', 'born_usa_f', 'party_f', 'religstr_f', 'reltrad_f', 'region_f', 'realinc2004_f')
clean_f_controls <- str_remove(f_controls, "_f$")

pca_data <- data2004 %>%
  select(all_of(c(n_outcomes, f_controls))) %>%
  rename_with(~ str_remove(., "_f$")) %>% # Remove _f from variables names for cleaner outputs
  rename_with(~ str_remove(., "_n$"))

pca_results <- PCA(pca_data, scale.unit = TRUE, quali.sup = clean_f_controls, ncp = 10, graph = FALSE)
```

```{r}
# Fonction pour tracer les corrplots
draw_corrplot <- function(data, title, subtitle, filename, range) {
  corrplot <- ggcorrplot(
    data,
    method = "square",
    ggtheme = theme_dark,
    show.legend = TRUE,
    outline.color = "#636363",
    tl.cex = 8) +
    
    theme(
      legend.title = element_blank(),
      legend.key.width = unit(0.3, "cm"),
      legend.key.height = unit(1, "null")) +
    
    if (identical(range, c(-1,1))) {
      scale_fill_gradient2(limit = c(-1,1), 
                           low = "#6D9EC1", 
                           mid = "white", 
                           high =  "#E46726", 
                           midpoint = 0, 
                           oob = scales::oob_squish)}
    else if (identical(range, c(0,1))) {
      scale_fill_gradient2(limit = c(0, 1), 
                           low = "white", 
                           high = "#E46726", 
                           oob = scales::oob_squish)}
  
  # On ajoute titres et sous-titres pour les graphs affichés
  corrplot <- corrplot + 
    ggtitle(title) +
    theme(plot.title = element_text(hjust = 0.5, size = 12))
  
  if (subtitle != "") {
    corrplot <- corrplot + 
      labs(subtitle = subtitle) + 
      theme(plot.subtitle = element_text(hjust = 0.5, size = 12))}
  
  ggsave(filename, plot = corrplot)
  
  print(corrplot)
}
```

```{r}
## Histogramme de la contribution des axes à la variance totale
screeplot <- fviz_screeplot(pca_results, addlabels = TRUE, ylim = c(0, 45), title = "") + theme_gray()
screeplot$layers[[1]]$aes_params$fill <- "lightskyblue"
screeplot$layers[[1]]$aes_params$colour <- "cornflowerblue"
screeplot$layers[[2]]$aes_params$colour <- "darkblue"
screeplot$layers[[3]]$aes_params$colour <- "darkblue"
screeplot$layers[[4]]$aes_params$colour <- "darkblue"
screeplot$labels$x <- "Components"
screeplot$labels$y <- "% of explained variance"
screeplot$labels$title = "Screeplot"
screeplot$theme$plot.title$hjust <- 0.5
screeplot
ggsave("output/acp/screeplot.png", plot = screeplot, width = 10, height = 6, units = "in", dpi = 300)
```

```{r}
## Contribution des variables aux axes
## Correspond à la qualité de représentation des variables (= cos2) normalisée entre 1 et 100
## Présentable à l'aide d'un tableau...
coeffs <- pca_results$var$contrib
labels <- rownames(pca_results$var$contrib)
axis_df <- data.frame(labels, coeffs, row.names = NULL)
colnames(axis_df) <- c("Variables", "1", "2", "3", "4", "5", "6", "7", "8", "9", "10")
axis_tbl <- axis_df %>%
  gt() %>%
  fmt_number(decimals = 2, drop_trailing_zeros = TRUE) %>%
  tab_header("Variable contributions to the components’ variance", subtitle = NULL, preheader = NULL) %>%
  tab_spanner(label = "Components", columns = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10"), id = "axes") %>%
  data_color(method = "numeric", columns = !matches("Variables"), palette ="Oranges", direction = "column") %>%
  tab_row_group(label = "How close do you feel to?", rows = which(.data$Variables %in% q2_vars), id = "close") %>%
  tab_row_group(label = "How important do you think each of the following is?", rows = which(.data$Variables %in% q3_vars), id = "imp") %>%
  tab_row_group(label = "How much do you agree or disagree with the following statements?", rows = which(.data$Variables %in% q4_vars), id = "agree") %>%
  tab_row_group(label = "How proud are you of America in each of the following?", rows = which(.data$Variables %in% q5_vars), id = "proud") %>%
  row_group_order(c("close", "imp", "agree", "proud")) %>%
  tab_style(
    style = cell_text(align = "center"),
    locations = list(
      cells_column_labels(columns = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10")),
      cells_body(columns = !matches("Variables")))) %>%
  tab_style(
    style = cell_fill(color = "lightgrey"),
    locations = cells_column_labels(columns = c("1", "3", "5", "7", "9"))) 
axis_tbl
gtsave(axis_tbl, filename = "contrib_axes.docx", path = "output/acp/")

## ... d'un corrplot ...
draw_corrplot(t(pca_results$var$contrib[, 1:10]/100), "Variable contributions to the components’ variance", "", "output/acp/contrib_axes.png", range = c(0,1))

## ... ou de graphs
contrib_1 <- fviz_contrib(pca_results, choice = "var", axes = 1, top = 30) + 
  theme_gray() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
contrib_1$layers[[1]]$aes_params$fill <- "lightskyblue"
contrib_1$layers[[1]]$aes_params$colour <- "cornflowerblue"
contrib_1$layers[[2]]$aes_params$colour <- "darkblue"
contrib_1$labels$title <- "Variable contributions to the 1st component variance"
contrib_1$theme$plot.title$hjust <- 0.5
contrib_1
ggsave("output/acp/contrib_axe_1.png", plot = contrib_1)

contrib_2 <- fviz_contrib(pca_results, choice = "var", axes = 2, top = 30) + 
  theme_gray() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
contrib_2$layers[[1]]$aes_params$fill <- "lightskyblue"
contrib_2$layers[[1]]$aes_params$colour <- "cornflowerblue"
contrib_2$layers[[2]]$aes_params$colour <- "darkblue"
contrib_2$labels$title <- "Variable contributions to the 2nd component variance"
contrib_2$theme$plot.title$hjust <- 0.5
contrib_2
ggsave("output/acp/contrib_axe_2.png", plot = contrib_2)
```

```{r}
## Corrélation des variables aux axes
## Avec à nouveau deux représentations possibles...
## Des corrplots
draw_corrplot(t(pca_results$var$coord[, 1:5]), title = "Correlation of outcome variables with the components", "", "output/acp/corr_outc_axes.png", range = c(-1,1))
draw_corrplot(t(pca_results$quali.sup$coord[, 1:5]), "Correlation of control variables", "with the components", "output/acp/corr_rég_axes.png", range = c(-1,1))
## Des projections sur les plans factoriels
proj_var <- fviz_pca_var(pca_results, 
                         axes = c(1, 2),
                         select.var = list(contrib = 12),
                         col.circle = '#636363',
                         col.var = "contrib",
                         gradient.cols = c("cornflowerblue", "#E46726"),
                         title = "Variable projections on components 1 and 2", 
                         repel = TRUE) + 
  theme_gray() +
  theme(plot.title = element_text(hjust = 0.5))
proj_var$layers <- proj_var$layers[-c(4, 5)] # removing original hline and vline
proj_var <- proj_var +
  geom_hline(yintercept = 0, color = "#636363", alpha = 0.5, linetype = "twodash") +
  geom_vline(xintercept = 0, color = "#636363", alpha = 0.5, linetype = "twodash") + 
  theme_gray() +
  labs(x = paste0("Component 1 (", round(screeplot$data$eig[1], 2), "%)")) +
  labs(y = paste0("Component 2 (", round(screeplot$data$eig[2], 2), "%)"))
proj_var
ggsave("output/acp/proj_var.png", plot = proj_var)
```

```{r}
proj_ind <- fviz_pca_ind(pca_results, 
                         axes = c(1, 2),
                         habillage = "party",
                         pointshape = 16,
                         alpha.ind = 0.7,
                         title = "Individual projections on components 1 and 2", 
                         label = "none",
                         repel = TRUE) +
  theme_gray() +
  scale_color_manual(
    name = "Partisanship",
    values = c("Republican" = "coral1", "Independent" = "lightgreen", "Democrat" = "cornflowerblue")) +
  theme(plot.title = element_text(hjust = 0.5))
proj_ind$layers <- proj_ind$layers[-c(2, 3)] # removing original hline and vline
proj_ind <- proj_ind +
  geom_hline(yintercept = 0, color = "#636363", linetype = "twodash") +
  geom_vline(xintercept = 0, color = "#636363", linetype = "twodash") +
  labs(x = paste0("Component 1 (", round(screeplot$data$eig[1], 2), "%)")) +
  labs(y = paste0("Component 2 (", round(screeplot$data$eig[2], 2), "%)"))
proj_ind
ggsave("output/acp/proj_ind.png", plot = proj_ind)

proj_ind <- proj_ind + xlim(-3, 3) + ylim(-3, 3) # Zoom in
ggsave("output/acp/proj_ind_zoom.png", plot = proj_ind)
```

```{r}
## Qualité de représentation des variables
draw_corrplot(t(pca_results$var$cos2[, 1:5]), "Quality of outcomes variables", "representation", "output/acp/qual_rep_outc.png", range = c(0,1))
draw_corrplot(t(pca_results$quali.sup$cos2[, 1:5]), "Quality of control variables", "representation", "output/acp/qual_rep_régr.png", range = c(0,1))
```
